+++ src/client/java/com/skycatdev/autocut/config/ExportGroupingMode.java
@@ -8,19 +8,19 @@ import org.jetbrains.annotations.Nullable;
 /**
  * What group a clip should be exported with
  */
-public enum ExportGroupingMode { // TODO: Localize
+public enum ExportGroupingMode {
     /**
      * Keep it with all the other clips
      */
-    NONE(Identifier.of(Autocut.MOD_ID, "none"), Text.of("Main file")),
+    NONE(Identifier.of(Autocut.MOD_ID, "none"), Text.translatable("autocut.exportGroupingMode.none")),
     /**
      * Put it in a file with the things of its type
      */
-    TYPE(Identifier.of(Autocut.MOD_ID, "type"), Text.of("Own type")),
+    TYPE(Identifier.of(Autocut.MOD_ID, "type"), Text.translatable("autocut.exportGroupingMode.type")),
     /**
      * Give it its own file
      */
-    INDIVIDUAL(Identifier.of(Autocut.MOD_ID, "individual"), Text.of("Separate"));
+    INDIVIDUAL(Identifier.of(Autocut.MOD_ID, "individual"), Text.translatable("autocut.exportGroupingMode.individual"));
 
     private final Identifier id;
     private final Text name;